stage resolver
