+++ src/flow/flow/flow.graphql.models.ts
@@ -1,26 +1,7 @@
 import { Field, InputType, Int, ObjectType } from '@nestjs/graphql';
-import { Task } from '../task/task.graphql.models';
+import {Stage} from '../stagedb/stage.graphql.model'
 
-@ObjectType()
-export class Stage {
-    @Field(type => String, { nullable: false })
-    _id: string;
 
-    @Field(type => [Task], { nullable: false, description: "list of tasks in the stage" })
-    tasks: Task[];
-
-    @Field(type => String, { nullable: false, description: "provide a unique name" })
-    name?: string;
-
-    @Field(type => String)
-    createdDate?: string;
-
-    @Field(type => Boolean, { nullable: true, description: "can the customer skip the flow" })
-    canSkip: boolean;
-
-    @Field(type => Boolean, { nullable: true, description: "does it need host Consent To Proceed the flow" })
-    hostConsentToProceed: boolean;
-}
 
 @ObjectType()
 export class Flow {
