+++ src/flow/stagedb/stage.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, InputType, Mutation, Parent, Query, ResolveField, Resolver, Subscription } from "@nestjs/graphql";
-import {StagedbService} from './stagedb.service'
-import {Stage, StageInput} from './stage.graphql.model'
+import { StagedbService } from './stagedb.service'
+import { Stage, StageInput } from './stage.graphql.model'
 
 @Resolver(of => Stage)
 export class StageResolver {
@@ -8,14 +8,21 @@ export class StageResolver {
         private stageDbService: StagedbService
     ) { }
 
-    @Query(returns => Stage)
-    async getStageById( @Args('id', { type: () => String }) id:string){
+    @Query(returns => Stage, { description: "returns a Stage with a given ID" })
+    async getStageById(@Args('id', { type: () => String }) id: string) {
         return this.stageDbService.getById(id)
-    } 
+    }
 
-    @Query(returns => [Stage])
-    async getAllStages(){
+    @Query(returns => [Stage], { description: "returns all Stages" })
+    async getAllStages() {
         return this.stageDbService.getAll()
     }
+
+    @Mutation(returns => Stage, { description: "creates a Stage and returns the response" })
+    async createStage(@Args('Stage', { type: () => StageInput }) Stage: StageInput) {
+        return this.stageDbService.save(Stage)
+    }
+
+
 
 }
